Gold & platinum rebalance
A lot of dicking around with Gold to up the fabber count while still
having it expand slower than Plat and Uber. Didn't work so we're back to
the old fabber count with just the troop rebalances left. Experimenting
with Plat over fabbering, but need to tune this better for naval

Gold is less sophisticated in its troop choices
Gold is slower to react to mines
Gold will build the Slammer
Gold through Uber use less Dox scouts on smaller planets